requirements.txt
# Core dependencies
aiohttp>=3.8.0
pyyaml>=6.0
readability-lxml>=0.8.1
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Optional: For better HTML parsing
html5lib>=1.1